requirements.txt
instabot==0.32.0
python-dotenv==0.10.1